pyproject.toml
[tool.poetry]
name = "mortgagelen"
version = "0.1.0"
description = ""
authors = ["Mihail Butnaru <butnarum@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
py-dateutil = "^2.2"
pydantic = "^1.7.3"
black = "^20.8b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
